feat: reject overlapping put
The default behaviour of avltree.Put() is to overwrite an existing node
when the comparator returns an equal value. We don't want to do this
with intervals, because an existing interval will represent a
booking. This is not user-definable behaviour in emirpasic/gods library,
nor can it be made backwards compatible, because there it requires a
return from the function that indicates if the put was accepted. Hence
the avltree code is modified.