interval(fix): remove admin lock code from methods
We use the same methods for admin and user actions, e.g. making bookings
either as a user or as an admin replacing bookings. So we can't have the
methods throwing an error if the store is locked to users, because then
admins can't do the work they intended to do while the store was locked