serve(fix): remove body for CancelBooking
We simplify life for the consuming javascript by omitting the need to
include the booking in the body. It does not add any additional security
because anyone who can cancel a users bookings can also obtain the full
details of the bookings necessary to make the cancellation. The small
additional inconvenience does not constitute greater security. To
implement this more efficiently, a new function was added to store,
which can get a booking by name, so that the CancelBooking function can
be left in its current form with the argument being a complete booking