server (fix): intermittent fail (TestLockedToUser)
Not re-setting time at the start of each mini test, meant that
sometimes the bookings would be uploaded after another copy of
GetActivity had moved time beyond the start of the booking that
we wanted the Activity for. Fixed mini test setup code to set
the time before uploading the bookings.

[Issue: 10]