cmd (fix): report build version & time
The build.sh script was not setting the Version and BuildTime flags
because there were still references to the old timdrysdale/interval
package, so they were getting set in that package instead. Updating all
the internal packages to exclusively refer to this module, as it should
be, has fixed it.