api (feat): add groups, fix warnings
Groups were added to the API as the primary means of allocating policies to users,
requiring addition of two new endpoints, and removal of two
endpoints (no need to get policies, or add policies to user account
now). Warnings for GetActivity were fixed (bad example was
removed). CheckManifest now returns 200 OK with an error list if the
manifest is succesfully checked but contains errors. REST status codes
should be successful if request is successful, and do not represent
function return values for whatever the request was. Internal error
would only occur if the manifest could not actually be checked. On
uploading, an incorrect manifest still results in 500 error though, because it
caused an internal server issue in trying to process the manifest. TODO
- update server code to implement API changes, and update tests.

[Issue: 13]